Add IAbpClaimsIdentityService
