+++ framework/src/Volo.Abp.Security/Volo/Abp/Security/Claims/ClaimsIdentityContext.cs
@@ -1,17 +1,20 @@
 ﻿using System;
 using System.Security.Claims;
+using JetBrains.Annotations;
 
 namespace Volo.Abp.Security.Claims
 {
     public class ClaimsIdentityContext
     {
+        [NotNull]
         public ClaimsIdentity ClaimsIdentity { get; }
 
+        [NotNull]
         public IServiceProvider ServiceProvider { get; }
 
         public ClaimsIdentityContext(
-            ClaimsIdentity claimsIdentity,
-            IServiceProvider serviceProvider)
+            [NotNull] ClaimsIdentity claimsIdentity,
+            [NotNull] IServiceProvider serviceProvider)
         {
             ClaimsIdentity = claimsIdentity;
             ServiceProvider = serviceProvider;
