+++ framework/src/Volo.Abp.Security/Volo/Abp/Security/Claims/IAbpClaimsIdentityService.cs
@@ -1,0 +1,10 @@
+﻿using System.Security.Claims;
+using System.Threading.Tasks;
+
+namespace Volo.Abp.Security.Claims
+{
+    public interface IAbpClaimsIdentityService
+    {
+        Task AddClaimsAsync(ClaimsIdentity identity);
+    }
+}
